move stopwatch component into a timers folder and update links
